requirements.txt
pandas==1.1.3
numpy==1.21.0
requests==2.28.1
wandb==0.13.2
streamlit==1.12.2
torchvision==0.10.0
torch==1.9.0
cloudpickle==1.6.0
Pillow==9.2.0
rsbox==0.0.7
torchplate==0.0.6
